refactor: slot 분리
