+++ src/main/java/com/onebyte/life4cut/slot/domain/Slot.java
@@ -1,8 +1,9 @@
-package com.onebyte.life4cut.album.domain;
+package com.onebyte.life4cut.slot.domain;
 
-import com.onebyte.life4cut.album.domain.vo.SlotLayout;
-import com.onebyte.life4cut.album.domain.vo.SlotLocation;
+import com.onebyte.life4cut.album.domain.Album;
 import com.onebyte.life4cut.common.entity.BaseEntity;
+import com.onebyte.life4cut.slot.domain.vo.SlotLayout;
+import com.onebyte.life4cut.slot.domain.vo.SlotLocation;
 import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
 import jakarta.persistence.Column;
