+++ src/test/java/com/onebyte/life4cut/picture/service/PictureServiceIntTest.java
@@ -11,7 +11,6 @@ import com.onebyte.life4cut.album.domain.vo.UserAlbumRole;
 import com.onebyte.life4cut.album.exception.AlbumNotFoundException;
 import com.onebyte.life4cut.album.exception.UserAlbumRolePermissionException;
 import com.onebyte.life4cut.album.repository.AlbumRepositoryImpl;
-import com.onebyte.life4cut.album.repository.SlotRepositoryImpl;
 import com.onebyte.life4cut.album.repository.UserAlbumRepositoryImpl;
 import com.onebyte.life4cut.common.constants.S3Env;
 import com.onebyte.life4cut.config.JpaConfiguration;
@@ -28,6 +27,7 @@ import com.onebyte.life4cut.picture.exception.PictureNotFoundException;
 import com.onebyte.life4cut.picture.repository.PictureRepositoryImpl;
 import com.onebyte.life4cut.picture.repository.PictureTagRelationRepositoryImpl;
 import com.onebyte.life4cut.pictureTag.repository.PictureTagRepositoryImpl;
+import com.onebyte.life4cut.slot.repository.SlotRepositoryImpl;
 import com.onebyte.life4cut.support.fileUpload.FileUploadResponse;
 import com.onebyte.life4cut.support.fileUpload.FileUploader;
 import com.querydsl.jpa.impl.JPAQueryFactory;
