+++ src/test/java/com/onebyte/life4cut/picture/service/PictureServiceTest.java
@@ -8,14 +8,11 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.onebyte.life4cut.album.domain.Slot;
 import com.onebyte.life4cut.album.domain.vo.UserAlbumRole;
 import com.onebyte.life4cut.album.exception.AlbumDoesNotHaveSlotException;
 import com.onebyte.life4cut.album.exception.AlbumNotFoundException;
-import com.onebyte.life4cut.album.exception.SlotNotFoundException;
 import com.onebyte.life4cut.album.exception.UserAlbumRolePermissionException;
 import com.onebyte.life4cut.album.repository.AlbumRepository;
-import com.onebyte.life4cut.album.repository.SlotRepository;
 import com.onebyte.life4cut.album.repository.UserAlbumRepository;
 import com.onebyte.life4cut.common.constants.S3Env;
 import com.onebyte.life4cut.fixture.AlbumFixtureFactory;
@@ -29,6 +26,9 @@ import com.onebyte.life4cut.picture.domain.vo.PictureTagName;
 import com.onebyte.life4cut.picture.repository.PictureRepository;
 import com.onebyte.life4cut.picture.repository.PictureTagRelationRepository;
 import com.onebyte.life4cut.pictureTag.repository.PictureTagRepository;
+import com.onebyte.life4cut.slot.domain.Slot;
+import com.onebyte.life4cut.slot.exception.SlotNotFoundException;
+import com.onebyte.life4cut.slot.repository.SlotRepository;
 import com.onebyte.life4cut.support.fileUpload.FileUploadResponse;
 import com.onebyte.life4cut.support.fileUpload.FileUploader;
 import java.time.LocalDateTime;
